pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "json file differences generator"
authors = ["jkulds <w84my5@yandex.ru>"]
readme = "README.md"
packages = [{include = "gendiff"}]
repository = "https://github.com/jkulds/python-project-50"
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = ">=7.4.0"
pytest-cov = ">=4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
